+++ src/app/components/Pages/DASHBOARD/user-btn.jsx
@@ -1,7 +1,7 @@
 import { Avatar, AvatarImage } from '@/components/ui/avatar';
 import { Button } from '@/components/ui/button';
 
-import { auth } from '@/lib/auth';
+import { auth } from '@/lib/auth/auth';
 
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuLabel, DropdownMenuTrigger } from '@/components/ui/dropdown-menu';
 import { SignOut } from './SignOut';