+++ src/middleware.js
@@ -1,5 +1,5 @@
 import { NextResponse } from 'next/server';
-import { auth } from './lib/auth';
+import { auth } from './lib/auth/auth';
 
 export async function middleware(req) {
   const session = await auth();