+++ CHANGELOG.md
@@ -4,10 +4,9 @@ All notable changes to this project will be documented in this file. See [standa
 
 ### [0.17.4](https://github.com/akshayfanatic/weelp_frontend/compare/v0.17.3...v0.17.4) (2025-09-30)
 
-
 ### 🐛 Bug Fixes
 
-* **ProductCard:** Fix Margin of Product Card According To Design ([b2465ef](https://github.com/akshayfanatic/weelp_frontend/commit/b2465ef655c521543ca4661f304d315f6f3fede1))
+- **ProductCard:** Fix Margin of Product Card According To Design ([b2465ef](https://github.com/akshayfanatic/weelp_frontend/commit/b2465ef655c521543ca4661f304d315f6f3fede1))
 
 ### [0.17.3](https://github.com/akshayfanatic/weelp_frontend/compare/v0.17.2...v0.17.3) (2025-09-29)